app/requirements.txt
eval_type_backport==0.3.1
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
psycopg[binary]==3.2.13
python-dotenv==1.0.1
pydantic==2.10.3
openpyxl==3.1.5
